requirements.txt
PyMySQL==1.1.0
PyYAML==6.0.1
telebot==0.0.5
schedule==1.2.1
minio==7.2.5

